fix: APIBackendListenerClient优化
